+++ isaaclab_arena_examples/relations/isaac_sim_kitchen_background_collision_notebook.py
@@ -53,8 +53,8 @@ def run_kitchen_background_collision_demo(simulation_app, view_steps: int = 0, a
     from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
-    from isaaclab_arena.relations.background_colliders import build_placement_region, find_background_colliders
     from isaaclab_arena.relations.relations import IsAnchor, On
+    from isaaclab_arena.scene.background_colliders import build_placement_region, find_background_colliders
     from isaaclab_arena.scene.scene import Scene
 
     if args_cli is None:
